requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cssselect==1.3.0
cssutils==2.11.1
Flask==3.1.2
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
gunicorn==20.1.0
httplib2==0.31.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
more-itertools==10.8.0
numpy==2.3.3
oauth2client==4.1.3
oauthlib==3.3.1
pandas==2.3.2
premailer==3.10.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yagmail==0.15.293
